refactor(proxy-capture): isolate capture session ownership
